pyproject.toml
[project]
name = "mnemox-lite"
version = "0.1.0"
description = "Semantic memory service for LLMs with modern GUI"
authors = [
    {name = "MnemoX Team"}
]
dependencies = [
    "qdrant-client>=1.7.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0", 
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0,<1.0.0",
    "sqlite-utils>=3.34.0",
    "mcp>=1.0.0",
    "google-genai>=1.0.0",
    "customtkinter>=5.2.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "psutil>=5.9.0"
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
